+++ documentation/Changelog.md
@@ -8,6 +8,7 @@
 ---
 ## 1.4.3
 Not released 
+* [CONJ-284] Cannot read autoincremented IDs bigger than Short.MAX_VALUE
 * [CONJ-282] Handling YEARs with binary prepareStatement
 * [CONJ-278] Improve prepared statement on failover